refactor: Phase 11.2 - Extract profile page components
Extract profile.vue into focused components:
- ProfileAccountInfo (71 lines) - Account details and tier badge
- ProfilePersonalInfo (111 lines) - Profile form for name/display name
- ProfileSecurity (129 lines) - Password change form
- ProfileDataPrivacy (62 lines) - Data export functionality
- ProfileDangerZone (109 lines) - Sign out and account deletion

Reduces profile.vue from 551 → 173 lines (-68.6%)
Better separation of concerns, easier testing and maintenance
